fix: update other src imports for move
