+++ src/student_success_tool/targets/pdp/failure_to_earn_enough_credits_in_time_from_enrollment.py
@@ -5,7 +5,8 @@ from collections.abc import Collection
 import numpy as np
 import pandas as pd
 
-from ... import features, utils
+from ... import utils
+from ...preprocessing import features
 from ...preprocessing.pdp import dataops
 from . import shared
